wip: improve type stability and simplifications
I forgot there was a lot of added complexity with the vectorization
experiment, so let's remove it from the frep definitions.

This also attempt to fix outstanding type stability issues in
CSG ops by using vectors for the RHS. See: examples/2d.jl

There are some regressions in FRep and other areas that need
to be sorted out.